Fix start script (#961)
