+++ tools/start.js
@@ -2,6 +2,7 @@ var webpack = require('webpack');
 var webpackConfig = require('../webpack.config');
 var webpackDevServer = require('webpack-dev-server');
 var detect = require('detect-port');
+var processConstEnum = require('./buildTools/processConstEnum');
 
 function startWebpackDevServer() {
     port = webpackConfig.devServer.port;
@@ -23,4 +24,5 @@ function startWebpackDevServer() {
         });
 }
 
+processConstEnum();
 +startWebpackDevServer();
